Changed LessonsStep
