+++ frontend/src/pages/CourseEditor/types.ts
@@ -1,4 +1,4 @@
-import { ChangeEventHandler } from 'react';
+import { ChangeEventHandler, BaseSyntheticEvent } from 'react';
 
 import { TCourseTechnology, ICourse } from 'types/course';
 
@@ -9,6 +9,10 @@ export interface IFormik {
     complexity?: number;
     avatar?: string;
     description?: string;
+    videoURL?: string;
+    textDescription?: string;
+    exerciseTitle?: string;
+    exerciseDescription?: string;
   };
   values: {
     technologies: TCourseTechnology[];
@@ -16,6 +20,10 @@ export interface IFormik {
     complexity?: number;
     avatar?: string;
     description?: string;
+    videoURL?: string;
+    textDescription?: string;
+    exerciseTitle?: string;
+    exerciseDescription?: string;
   };
   handleChange: ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement>;
 }
@@ -34,4 +42,23 @@ export interface ISkillsStepProps extends IStepProps {
   ungroupedSkills: {
     [key: string]: { group: string; image: string; maxScore: number; name: string; _id: string };
   };
+}
+
+export interface ILessonsStepContainerProps {
+  formik: IFormik;
+}
+
+export interface ILessonsStepProps {
+  formik: IFormik;
+  selectOption: string;
+  handleBackButtonClick?: (event: BaseSyntheticEvent) => void;
+  handleChangeOption?: (event: BaseSyntheticEvent) => void;
+  stageNext?: () => void;
+  stageBack?: () => void;
+}
+
+export interface ILessonItemProps {
+  formik: IFormik;
+  selectOption: string;
+  handleChangeOption?: (event: BaseSyntheticEvent) => void;
 }
